requirements.txt
# Windows Key Remapper Dependencies

# GUI Framework (modern-looking tkinter)
customtkinter>=5.2.0

# System tray support
pystray>=0.19.0
Pillow>=10.0.0

# For building executable
pyinstaller>=6.0.0
